Move tests to separate file
